fix do not clean personal tmp on launching vscode

+++ home/.bash_profile
@@ -1,16 +1,30 @@
+if ! grep -q file:///home/$USER/Pictures/Screenshots  ~/.config/gtk-3.0/bookmarks; then
+  echo file:///home/$USER/Pictures/Screenshots >> ~/.config/gtk-3.0/bookmarks
+fi
+
+if ! grep -q file:///home/$USER/Pictures/Screenshots  ~/.config/gtk-3.0/bookmarks; then
+  echo file:///home/$USER/Pictures/Screenshots >> ~/.config/gtk-3.0/bookmarks
+fi
 
 if [ ! -d ~/.tmp ]; then
   mkdir -p ~/.tmp
 fi
 
-rm -rf ~/.tmp/*
-
 if ! grep -q file:///home/$USER/.tmp  ~/.config/gtk-3.0/bookmarks; then
   echo file:///home/$USER/.tmp TMP >> ~/.config/gtk-3.0/bookmarks
 fi
 
-if ! grep -q file:///home/$USER/Pictures/Screenshots  ~/.config/gtk-3.0/bookmarks; then
-  echo file:///home/$USER/Pictures/Screenshots >> ~/.config/gtk-3.0/bookmarks
+#  env | sort > ~/bashrc-env.txt
+if [ -z $VSCODE_PID ]; then
+  rm -rf ~/.tmp/*
+fi
+
+if [ ! -d ~/Documents/TMP-MANUAL ]; then
+  mkdir -p ~/Documents/TMP-MANUAL
+fi
+
+if ! grep -q file:///home/$USER/Documents/TMP-MANUAL  ~/.config/gtk-3.0/bookmarks; then
+  echo file:///home/$USER/Documents/TMP-MANUAL >> ~/.config/gtk-3.0/bookmarks
 fi
 
 # SAVE_DIR="/home/$USER/afolder/_save/$(date +%Y-%m)-01"
